requirements.txt
Flask==3.0.2
google-cloud-storage==2.14.0
google-cloud-pubsub==2.19.0
python-dotenv==1.0.1
Werkzeug==3.0.1